Convert minehunter to new thread controler
